Add solutions for Lecture 1
Explicitly setting the Integer type for literals, such as 2 and 0, feels
a bit overkill. Then again, if the compiler warns about doing it itself
if we do not.

Use `let ... in` to make the recursive call easier to understand.
Still believe `lowerCount` and `greaterCount` could be improved
further.
In general `lowerAndGreater` seems a bit clunky and definitely not
elegant. Not sure how to keep lines shorter than 80 chars either. Maybe
to revisit.